Enable/disable of transformations added

+++ ckanext/extractor/controller.py
@@ -15,13 +15,15 @@ log = getLogger(__name__)
 
 class ExtractorController(PackageController):
 
-    def get_transformation(self,id):
+    def get_transformation_data(self,id):
         transformation = model.Session.query(Transformation).filter_by(package_name=id).first()
+        print transformation
 
         if transformation is not None:
             c.timestamp = transformation.timestamp.isoformat()
             c.mainclass = transformation.mainclass
             c.filename = transformation.filename
+            c.enabled = transformation.enabled
             c.data = True
         else:
             c.data = False
@@ -34,7 +36,7 @@ class ExtractorController(PackageController):
         # using default functionality
         template = self.read(id)
 
-        self.get_transformation(id)
+        self.get_transformation_data(id)
 
         c.error = False
 
@@ -47,47 +49,60 @@ class ExtractorController(PackageController):
          # using default functionality
         template = self.read(id)
 
-        filename = request.params['transformation_code'].filename
-        submitted_file = request.params['transformation_code'].file
-        mainclass = request.params['mainclass']
+        transformation = model.Session.query(Transformation).filter_by(package_name=id).first()
+        if transformation is None:
+            transformation = Transformation(id)
 
-        try:
-            #create transformations directory if it does not exist
-            if not os.path.isdir('transformations'):
-                log.info('Creating transformations directory')
-                os.mkdir('transformations')
+        #read enabled status of transformation
+        transformation.enabled = 'enabled' in request.params
+            
 
-            #delete package directory if it exists
-            package_dir = os.path.join('transformations', id)
-            if os.path.isdir(package_dir):
-                log.info('Deleting package directory %s' % package_dir)
-                shutil.rmtree(package_dir)
+        #get mainclass value
+        if 'mainclass' in request.params:
+            transformation.mainclass = request.params['mainclass']
 
-            #create directory again and change to it
-            log.info('Creating package directory %s' % package_dir)
-            os.mkdir(package_dir)
-            os.chdir(package_dir)
+        #read submitted file
+        if 'transformation_code' in request.params and request.params['transformation_code'] is not u'':
+            try:
+                transformation.filename = request.params['transformation_code'].filename
+                submitted_file = request.params['transformation_code'].file
 
-            zipfile = ZipFile(submitted_file)
-            log.info('Extracting file %s to directory %s' % (filename, package_dir))
-            zipfile.extractall()
+                #create transformations directory if it does not exist
+                if not os.path.isdir('transformations'):
+                    log.info('Creating transformations directory')
+                    os.mkdir('transformations')
 
-            log.info('File %s extracted' % filename)
+                #delete package directory if it exists
+                package_dir = os.path.join('transformations', id)
+                if os.path.isdir(package_dir):
+                    log.info('Deleting package directory %s' % package_dir)
+                    shutil.rmtree(package_dir)
 
-            #insert data into database
-            data = submitted_file.read()
-            transformation = Transformation(id, filename, data, datetime.now(), mainclass)
+                #create directory again and change to it
+                log.info('Creating package directory %s' % package_dir)
+                os.mkdir(package_dir)
+                os.chdir(package_dir)
 
-            model.Session.merge(transformation)
-            model.Session.commit()
-            log.info("Transformation object stored for package '%s'" % id)
+                zipfile = ZipFile(submitted_file)
+                log.info('Extracting file %s to directory %s' % (transformation.filename, package_dir))
+                zipfile.extractall()
 
-            c.error = False
+                #obtain data
+                transformation.data = submitted_file.read()
+                transformation.timestamp = datetime.now()
 
-            self.get_transformation(id)
-        except Exception as e:
-            c.error = True
-            c.error_message = 'Problem handling uploaded file %s (%s)' % (filename, e)
+                log.info('File %s extracted' % transformation.filename)
+            except Exception as e:
+                c.error = True
+                c.error_message = 'Problem handling uploaded file %s (%s)' % (transformation.filename, e)
+                return render('extractor/read.html')
+
+        model.Session.merge(transformation)
+        model.Session.commit()
+        log.info("Transformation object stored for package '%s'" % id)
+
+        c = self.get_transformation_data(id)
+        c.error = False
 
         #rendering using default template
         return render('extractor/read.html')
